+++ libraries/strings/mod.ts
@@ -1,6 +1,8 @@
 export { Interpolable } from './src/Interpolable.ts';
 export { interpolate } from './src/interpolate.ts';
+export { obfuscate } from './src/obfuscate.ts';
 export { pickVariants } from './src/pickVariants.ts';
 export { slugify } from './src/slugify.ts';
+export { toCamelCase } from './src/toCamelCase.ts';
 export { unindent } from './src/unindent.ts';
 export { validateDelimiters } from './src/validateDelimiters.ts';